refactored to have a better package layout
